New: Adding the first command support
Engine supports the root node insertion, just need to be capable to
write a memory address.

Memory driver supports now read and write requests. Able to handle
different clocks for the RAM and the core, just need to introduce the
dedicated clock for the AXI RAM, the CDC paths are ready. Address and
data channels are passed thru dual clock FIFO for CDC purpose.

Updated the assertion in testbench by adding the signal name failling.
Deactivated some of them, no more relevant and not spec mandatory when
under reset or after reaching out.

Added a function to print testbench parameters configuring the core and
call it in setup() before each test execution.