New: Implement search command to retrieve a token payload
Extend bst_engine FSM to handle search. Return error is searching across
a NULL tree, or if searching for a token not inserted. Else return the
payload. While operations are executed in order, just forward the
payload and not its associated token value. bst_engine also not uses a
cpl interface moving back data to the interface handler.

Add several unit tests to check the expected behavior.

Start to code a single clock FIFO to put in place a memory driver using
the same clock than the AXI4 RAM. This will nleash performance and avoid
the dual clock fifo which introduces a lot of latency.