New: Support DELETE_TOKEN
Now BSTer core supports delete command. This new FSM relies in search
engine to locate the token to delete, or a new slot to insert an orphan
node after a parent deletion.

Further delete command will be available later, probably not in FSM
itself but in interface handler to program and sequence. For instance
delete a branch will be a sequence of search and delete instead of
overcharging the FSM.

Implement a logger at bst_engine level to print FSM states, interface
request and memory accesses.