New: Enhance tree space manager
The Tree Space Manager now embbed a SC-FIFO to store the token addresses
deleted, so is able to recycle them. If FIFO is not empty, the address
counter is not incremented and the module issues the address to recycle.
This is a first version, the FIFO can't be sized and its depth is big as
the RAM address width for the AXI4 interface. To enhance in the future
but not critical for the moment before moving to synthesis.